+++ CHANGELOG.md
@@ -244,3 +244,4 @@
 * Move ``Cowsay`` templates to a dedicated folders for bot assets.
 * Elegant event handler.
 * Help command is now useful.
+* Added ``realShuffle`` array utility method that uses a better prng (newly added ``crypto#randomInt``).